Move the transaction list into a component
